Refactored test runner
