+++ UnityProject/Assets/Tests/PerformanceTests/Utils/EditorRoutine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf698a9a6fe9aa640a4554319218a0ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
